A lot of changes + Pokedex
We've been attempting to implement a pokedex though it's not completely stable. Also, there's now a professor cutscene when you start the game, it'll ask you for your name and gender although it's not finalized. Custom sprites are halfway there, I can't seem to implement them correctly. Also the camera movement event will now allow for simultaneous movement. Oh, and you can now scroll dialogue, though it's not in the dialogue custom event yet. Once again, develop is entirely experimental and should not be used as your framework.