Wild Pokemon Encounter Rates Fix
The current implementation of Pokemon Encounter rates was global for the whole map. This is incorrect based on reading this post on pokemondb and various other sources: https://pokemondb.net/pokebase/254583/information-encounter-various-locations-description-carefully.

I edited PlayerMovement to correct interact with my changes in MapSettings, which was mainly to move Pokemon Rarity from being Map wide to Pokemon specific. If higher encounter rates are desired(even above 100%, no checks for that), just add more to the encounter list. I added comments to the enumeration of the Pokemon Rarirty to help quickly calculate rarity.

I did most of my prototyping in Overworld N and have left it there. This could be VERY useful in testing Pokemon Battle Sprites Front/Backs.