Research on Nintendo Rom Compared to PKU
there's an individual field for each berry plant in overworld
there's the daycare class missing...
a bool variable for if the pc background is unlocked
(some backgrounds are plot rewards)
also forgot to do the player savings variable
(player mom doubles as bank)
also forgot rival data
might even need to double it, in case there's multiple
i mean, array it, not double
the date player began adventure is different from duration played game...

`In B2W2 a pkm file has 180 fields. In BW pkm files had 176 fields.`
Pokemon Class in nintendo has a separate field for each value, and store a lot of information. Our class is a fraction of that when creating a pokemon with params requested...